[acpi] Def. get_acpi_init_status() as macro when MOD_acpi=0
Reason: only by defining get_acpi_init_status() as a function macro returning
ais_not_started, we can always benefit from dead-code-elimination like if (0).
That allows expressions like `if (get_acpi_init_status() >= XXX)` to act like
`if (MOD_acpi)` in all the cases (even with -O0 -fno-inline-functions). That
is super useful to avoid link errors with ACPI functions when the module is
not compiled it. In other words, it is safe to check get_acpi_init_status()
and then call ACPI functions: in case the module is compiled-in, it will work
as a regular function, in the other case it will trigger dead-code-elimination
and the calls to the ACPI functions won't be in the generated code anymore.